Reorganized components, deleted app.css
